lol stupid moment

+++ src/main/java/au/id/mcterra/logparser/Status.java
@@ -4,12 +4,12 @@ import org.json.simple.JSONObject;
 import org.json.simple.parser.JSONParser;
 
 public class Status {
-	public final int roundId;
+	public final long roundId;
 
 	public Status(String text) throws Exception {
 		JSONParser parser = new JSONParser();
 		Object result = parser.parse(text);
 		JSONObject json = (JSONObject)result;
-		roundId = (int)json.get("round_id");
+		roundId = (long)json.get("round_id");
 	}
 }
